Creating Resoponse messages and Body validations; Adding a .env example for JWT Token; Making some logic changes and translation in the files

+++ src/main/java/com/biblioteca/services/BookService.java
@@ -28,7 +28,13 @@ public class BookService {
         return bookRepository.save(book);
     }
 
-    public void deleteById(Long id) {
-        bookRepository.deleteById(id);
+    public String deleteById(Long id) {
+        Optional<BookEntity> book = bookRepository.findById(id);
+        if (book.isPresent()) {
+            bookRepository.deleteById(id);
+            return "Livro deletado com sucesso!";
+        } else {
+            return "Livro não encontrado!";
+        }
     }
 }
