+++ src/main/java/com/biblioteca/services/UserService.java
@@ -1,8 +1,12 @@
 package com.biblioteca.services;
 
 import java.util.List;
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import com.biblioteca.dto.EditUserDTO;
 import com.biblioteca.entities.UserEntity;
 import com.biblioteca.repositories.UserRepository;
 
@@ -11,12 +15,6 @@ public class UserService {
 
     @Autowired
     private UserRepository userRepository;
-
-    public UserEntity saveUser(UserEntity userEntity) {
-        String passEncoded = userEntity.getUserPassword();
-        userEntity.setUserPassword(passEncoded);
-        return userRepository.save(userEntity);
-    }
 
     public List<UserEntity> listAllUsers() {
         return userRepository.findAll();
@@ -27,22 +25,29 @@ public class UserService {
                 .orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
     }
 
-    public void deleteUser(Long userId) {
-        userRepository.deleteById(userId);
+    public String deleteUser(Long userId) {
+        Optional<UserEntity> user = userRepository.findById(userId);
+        if(user.isPresent()) {
+            userRepository.deleteById(userId);
+            return "Usuário deletado com sucesso!";
+        } else {
+            return "Usuário não encontrado!";
+        }
     }
 
-    public UserEntity updateUser(Long userId, UserEntity newUser) {
+    public UserEntity updateUser(Long userId, EditUserDTO newUser) {
         UserEntity existingUser = findUserById(userId);
-        existingUser.setUserName(newUser.getUsername());
-        existingUser.setUserCpf(newUser.getUserCpf());
 
-        if (newUser.getUserPassword() != null && !newUser.getUserPassword().isEmpty()) {
-            String passEncoded = newUser.getUserPassword();
+        existingUser.setUserName(newUser.name());
+        existingUser.setUserCpf(newUser.cpf());
+
+        if (newUser.password() != null && !newUser.password().isEmpty()) {
+            String passEncoded = newUser.password();
             existingUser.setUserPassword(passEncoded);
         }
 
-        existingUser.setUserEmail(newUser.getUserEmail());
-        existingUser.setUserType(newUser.getUserType());
+        existingUser.setUserEmail(newUser.email());
+        existingUser.setUserType(newUser.type());
         return userRepository.save(existingUser);
     }
 }
